refactor: Context API로 리팩토링
